修正错误的方法，修正错误的数据类型转换，合并FirstOrDefault与Where的调用

+++ src/Services/Services.Media/Core/Adapters/LiveAdapter.cs
@@ -31,7 +31,7 @@ internal static class LiveAdapter
     {
         var title = room.Title;
         var roomId = room.RoomId.ToString();
-        var viewerCount = System.Convert.ToDouble(room.ViewerCount);
+        var viewerCount = System.Convert.ToInt32(room.ViewerCount);
         var user = UserAdapterBase.CreateUserProfile(room.UserId ?? 0, room.UserName, room.UserAvatar, 48d);
         var cover = room.Cover.ToVideoCover();
         var subtitle = room.AreaName;
